merge master

+++ erupt-core/src/main/java/xyz/erupt/core/view/Page.java
@@ -8,8 +8,8 @@ import java.util.Collection;
 import java.util.Map;
 
 /**
- * @author liyuepeng
- * @date 2018-10-12.
+ * @author YuePeng
+ * date 2018-10-12.
  */
 @Getter
 @Setter
@@ -18,7 +18,7 @@ public class Page {
 
     private int pageSize;
 
-    private int totalPage;
+    private long totalPage;
 
     private Long total;
